# FastRegexMatcher: small optimization for the literal prefix case