In this PR I propose a small optimization in `FastRegexMatcher` for the literal prefix case. The idea in this PR is to split the actual prefix string matcher into 2: one of the case sensitive and one of the case insensitive case. It saves a couple of `if` condition checks.

Benchmark running on a subset of regexps:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
                                                     │  before.txt  │             after.txt              │
                                                     │    sec/op    │    sec/op     vs base              │
FastRegexMatcher/foo_.+-11                              71.12n ± 5%   70.86n ±  0%  -0.37% (p=0.002 n=6)
FastRegexMatcher/foo_.*-11                              70.87n ± 1%   70.86n ±  0%       ~ (p=0.483 n=6)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-11    771.6n ± 2%   773.5n ±  2%       ~ (p=0.589 n=6)
FastRegexMatcher/(?i:(AAAAAAAAAAAAAAAAAAAAAAAA|BB-11    320.5n ± 1%   324.3n ±  0%       ~ (p=0.327 n=6)
FastRegexMatcher/(?i:(zQPbMkNO.*|NNSPdvMi.*|iWuuS-11    5.208µ ± 5%   4.705µ ±  1%  -9.65% (p=0.002 n=6)
FastRegexMatcher/(?i:(.*zQPbMkNO|.*NNSPdvMi|.*iWu-11    6.164µ ± 0%   6.149µ ±  0%       ~ (p=0.320 n=6)
FastRegexMatcher/fo.?-11                                84.11n ± 1%   83.28n ±  1%  -0.99% (p=0.002 n=6)
FastRegexMatcher/foo.?-11                               83.42n ± 0%   82.84n ±  0%  -0.69% (p=0.002 n=6)
```